% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_utm.R
\name{find_utm}
\alias{find_utm}
\title{Find UTM zone of geographic coordinates}
\usage{
find_utm(lat, lon)
}
\arguments{
\item{lat}{Numeric vector. Latitude(s) in decimal degrees.}

\item{lon}{Numeric vector. Longitude(s) in decimal degrees.}
}
\value{
An integer vector of UTM zone numbers.
Note: Latitude bands (letters) are not included.
}
\description{
Given longitude and latitude coordinates in decimal degrees,
 returns the UTM zone number(s).
}
\examples{
find_utm(lat = 10, lon = -75)    # single coordinate
find_utm(lat = c(5, -10, 45), lon = c(-75, 140, 10))  # multiple coords

}
